## Summary
- **Faster room joins**: LiveKit token pre-warming fetches the token
while navigating, saving 1-2s
- **Soft update dialog**: Checks for updates on launch using existing
Firestore read (zero extra latency)
- **Broadcast banners**: Show quantity (3x Crown), restored cross-room
visibility, repositioned above seats
- **Super Shy purchase UX**: Sheet stays open during purchase instead of
showing blank loading screen
- **Self-destruct TTS**: Robotic voice announcement when room 5-min
countdown starts/stops (opt-in in Settings)
- **Large ShyBeans support**: Redemption uses Long throughout to handle
balances > 2.1 billion
- **In-room coin purchase**: Buy coins sheet available without leaving
the room
- **Invite fix**: Shows error message if the target user has left the
room
- **CF quantity**: lastGiftEvent includes quantity for all gift send
variants

## Test plan
- [x] 1698 Kotlin unit tests pass (including new selfDestructAlert,
invite, Long beans, User model tests)
- [x] 450 Cloud Functions tests pass
- [ ] Manual: Launch app → soft update dialog appears if update
available → dismiss works
- [ ] Manual: Tap room → joins noticeably faster (token pre-warmed)
- [ ] Manual: Send 3x gift → broadcast shows "3x Crown" in sender's room
and other rooms
- [ ] Manual: Settings > Notifications > Room Alerts > toggle
Self-Destruct Countdown
- [ ] Manual: Redeem All with large bean balance → succeeds
- [ ] Manual: Invite user who left → snackbar "X has left the room"

+++ app/src/androidTest/java/com/shyden/shytalk/fake/FakeEconomyRepository.kt
@@ -50,7 +50,7 @@ class FakeEconomyRepository : EconomyRepository {
     override suspend fun sendEntireBackpack(recipientId: String): Resource<Map<String, Any?>> =
         Resource.Success(mapOf("success" to true, "totalItemsSent" to 0))
 
-    override suspend fun redeemBeans(amount: Int): Resource<Map<String, Any?>> =
+    override suspend fun redeemBeans(amount: Long): Resource<Map<String, Any?>> =
         Resource.Success(mapOf("success" to true))
 
     override suspend fun purchaseCoins(productId: String, purchaseToken: String): Resource<Map<String, Any?>> =
